+++ src/model/resources/storage/storage_all.jl
@@ -9,6 +9,7 @@ function storage_all!(EP::Model, inputs::Dict, setup::Dict)
 
 	dfGen = inputs["dfGen"]
 	Reserves = setup["Reserves"]
+	capacity_reserve_margin = setup["CapacityReserveMargin"]
 
 	G = inputs["G"]     # Number of resources (generators, storage, DR, and DERs)
 	T = inputs["T"]     # Number of time steps (hours)
@@ -30,6 +31,15 @@ function storage_all!(EP::Model, inputs::Dict, setup::Dict)
 
 	# Energy withdrawn from grid by resource "y" at hour "t" [MWh] on zone "z"
 	@variable(EP, vCHARGE[y in STOR_ALL, t=1:T] >= 0);
+
+	# Virtual discharge contributing to capacity reserves at timestep t for storage cluster y
+	@variable(EP, vCAPRES_discharge[y in STOR_ALL, t=1:T] >= 0)
+
+	# Virtual charge contributing to capacity reserves at timestep t for storage cluster y
+	@variable(EP, vCAPRES_charge[y in STOR_ALL, t=1:T] >= 0)
+
+	# Total state of charge being held in reserve at timestep t for storage cluster y
+	@variable(EP, vCAPRES_socinreserve[y in STOR_ALL, t=1:T] >= 0)
 
 	### Expressions ###
 
@@ -89,14 +99,39 @@ function storage_all!(EP::Model, inputs::Dict, setup::Dict)
 		# Maximum discharging rate must be less than power rating OR available stored energy in the prior period, whichever is less
 		# wrapping from end of sample period to start of sample period for energy capacity constraint
 		@constraints(EP, begin
-			[y in STOR_ALL, t=1:T], EP[:vP][y,t] <= EP[:eTotalCap][y]
-			[y in STOR_ALL, t=1:T], EP[:vP][y,t] <= EP[:vS][y, hoursbefore(hours_per_subperiod,t,1)]*dfGen[y,:Eff_Down]
+			[y in STOR_ALL, t=1:T], EP[:vP][y,t] + EP[:vCAPRES_discharge][y,t] <= EP[:eTotalCap][y]
+			[y in STOR_ALL, t=1:T], EP[:vP][y,t] + EP[:vCAPRES_discharge][y,t] <= EP[:vS][y, hoursbefore(hours_per_subperiod,t,1)]*dfGen[y,:Eff_Down]
 		end)
 	end
 	#From co2 Policy module
 	@expression(EP, eELOSSByZone[z=1:Z],
 		sum(EP[:eELOSS][y] for y in intersect(STOR_ALL, dfGen[dfGen[!,:Zone].==z,:R_ID]))
 	)
+
+	# Capacity Reserve Margin policy
+	if capacity_reserve_margin == 1
+		# Constraints governing energy held in reserve when storage makes virtual capacity reserve margin contributions:
+
+		# Links energy held in reserve in first time step with decisions in last time step of each subperiod
+		# We use a modified formulation of this constraint (cVSoCBalLongDurationStorageStart) when operations wrapping and long duration storage are being modeled
+		@constraint(EP, cVSoCBalStart[t in START_SUBPERIODS, y in CONSTRAINTSET], EP[:vCAPRES_socinreserve][y,t] ==
+			EP[:vCAPRES_socinreserve][y,t+hours_per_subperiod-1] + (1/dfGen[y,:Eff_Down] * EP[:vCAPRES_discharge][y,t])
+			- (dfGen[y,:Eff_Up]*EP[:vCAPRES_charge][y,t]) - (dfGen[y,:Self_Disch] * EP[:vCAPRES_socinreserve][y,t+hours_per_subperiod-1]))
+
+		# energy held in reserve for the next hour
+		@constraint(EP, cVSoCBalInterior[t in INTERIOR_SUBPERIODS, y in STOR_ALL], EP[:vCAPRES_socinreserve][y,t] ==
+			EP[:vCAPRES_socinreserve][y,t-1]+(1/dfGen[y,:Eff_Down]*EP[:vCAPRES_discharge][y,t])-(dfGen[y,:Eff_Up]*EP[:vCAPRES_charge][y,t])-(dfGen[y,:Self_Disch]*EP[:vCAPRES_socinreserve][y,t-1]))
+		
+		# energy held in reserve acts as a lower bound on the total energy held in storage
+		@constraint(EP, cSOCMinCapRes[t in 1:T, y in STOR_ALL], EP[:vS][y,t] >= EP[:vCAPRES_socinreserve][y,t])
+	else
+		# Set values for all capacity reserve margin variables to 0
+		@constraints(EP, begin
+			[y in STOR_ALL, t in 1:T], EP[:vCAPRES_socinreserve][y,t] == 0
+			[y in STOR_ALL, t in 1:T], EP[:vCAPRES_discharge][y,t] == 0
+			[y in STOR_ALL, t in 1:T], EP[:vCAPRES_charge][y,t] == 0
+		end)
+	end
 end
 
 function storage_all_reserves!(EP::Model, inputs::Dict)
@@ -146,8 +181,8 @@ function storage_all_reserves!(EP::Model, inputs::Dict)
 
 			# Maximum discharging rate and contribution to reserves up must be less than power rating OR available stored energy in prior period, whichever is less
 			# wrapping from end of sample period to start of sample period for energy capacity constraint
-			[y in STOR_REG_RSV, t=1:T], EP[:vP][y,t]+EP[:vREG_discharge][y,t]+EP[:vRSV_discharge][y,t] <= EP[:eTotalCap][y]
-			[y in STOR_REG_RSV, t=1:T], (EP[:vP][y,t]+EP[:vREG_discharge][y,t]+EP[:vRSV_discharge][y,t])/dfGen[y,:Eff_Down] <= EP[:vS][y, hoursbefore(p,t,1)]
+			[y in STOR_REG_RSV, t=1:T], EP[:vP][y,t]+EP[:vCAPRES_discharge][y,t]+EP[:vREG_discharge][y,t]+EP[:vRSV_discharge][y,t] <= EP[:eTotalCap][y]
+			[y in STOR_REG_RSV, t=1:T], (EP[:vP][y,t]+EP[:vCAPRES_discharge][y,t]+EP[:vREG_discharge][y,t]+EP[:vRSV_discharge][y,t])/dfGen[y,:Eff_Down] <= EP[:vS][y, hoursbefore(p,t,1)]
 		end)
 	end
 	if !isempty(STOR_REG_ONLY)
@@ -174,8 +209,8 @@ function storage_all_reserves!(EP::Model, inputs::Dict)
 
 			# Maximum discharging rate and contribution to reserves up must be less than power rating OR available stored energy in prior period, whichever is less
 			# wrapping from end of sample period to start of sample period for energy capacity constraint
-			[y in STOR_REG_ONLY, t=1:T], EP[:vP][y,t] + EP[:vREG_discharge][y,t] <= EP[:eTotalCap][y]
-			[y in STOR_REG_ONLY, t=1:T], (EP[:vP][y,t]+EP[:vREG_discharge][y,t])/dfGen[y,:Eff_Down] <= EP[:vS][y, hoursbefore(p,t,1)]
+			[y in STOR_REG_ONLY, t=1:T], EP[:vP][y,t] + EP[:vCAPRES_discharge][y,t] + EP[:vREG_discharge][y,t] <= EP[:eTotalCap][y]
+			[y in STOR_REG_ONLY, t=1:T], (EP[:vP][y,t]+EP[:vCAPRES_discharge][y,t]+EP[:vREG_discharge][y,t])/dfGen[y,:Eff_Down] <= EP[:vS][y, hoursbefore(p,t,1)]
 		end)
 	end
 	if !isempty(STOR_RSV_ONLY)
@@ -196,16 +231,16 @@ function storage_all_reserves!(EP::Model, inputs::Dict)
 
 			# Maximum discharging rate and contribution to reserves up must be less than power rating OR available stored energy in prior period, whichever is less
 			# wrapping from end of sample period to start of sample period for energy capacity constraint
-			[y in STOR_RSV_ONLY, t=1:T], EP[:vP][y,t]+EP[:vRSV_discharge][y,t] <= EP[:eTotalCap][y]
-			[y in STOR_RSV_ONLY, t=1:T], (EP[:vP][y,t]+EP[:vRSV_discharge][y,t])/dfGen[y,:Eff_Down] <= EP[:vS][y, hoursbefore(p,t,1)]
+			[y in STOR_RSV_ONLY, t=1:T], EP[:vP][y,t]+EP[:vCAPRES_discharge][y,t]+EP[:vRSV_discharge][y,t] <= EP[:eTotalCap][y]
+			[y in STOR_RSV_ONLY, t=1:T], (EP[:vP][y,t]+EP[:vCAPRES_discharge][y,t]+EP[:vRSV_discharge][y,t])/dfGen[y,:Eff_Down] <= EP[:vS][y, hoursbefore(p,t,1)]
 		end)
 	end
 	if !isempty(STOR_NO_RES)
 		# Maximum discharging rate must be less than power rating OR available stored energy in prior period, whichever is less
 		# wrapping from end of sample period to start of sample period for energy capacity constraint
 		@constraints(EP, begin
-			[y in STOR_NO_RES, t=1:T], EP[:vP][y,t] <= EP[:eTotalCap][y]
-			[y in STOR_NO_RES, t=1:T], EP[:vP][y,t]/dfGen[y,:Eff_Down] <= EP[:vS][y, hoursbefore(p,t,1)]
+			[y in STOR_NO_RES, t=1:T], EP[:vP][y,t]  + EP[:vCAPRES_discharge][y,t] <= EP[:eTotalCap][y]
+			[y in STOR_NO_RES, t=1:T], (EP[:vP][y,t]+EP[:vCAPRES_discharge][y,t])/dfGen[y,:Eff_Down] <= EP[:vS][y, hoursbefore(p,t,1)]
 		end)
 	end
 end