+++ src/write_outputs/write_outputs.jl
@@ -48,6 +48,7 @@ function write_outputs(EP::Model, path::AbstractString, setup::Dict, inputs::Dic
 	dfCap = write_capacity(path, inputs, setup, EP)
 	dfPower = write_power(path, inputs, setup, EP)
 	dfCharge = write_charge(path, inputs, setup, EP)
+	dfVirtualDischarge = write_virtual_discharge(path, inputs, setup, EP)
 	dfCapacityfactor = write_capacityfactor(path, inputs, setup, EP)
 	elapsed_time_storage = @elapsed write_storage(path, inputs, setup, EP)
 	println("Time elapsed for writing storage is")